requirements.txt
itsdangerous==2.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==1.1.2
werkzeug==1.0.1
pymongo==4.0.1
greenlet==1.1.2
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
python-dotenv~=0.19.2
Flask-Cors==3.0.10
Flask-Script==2.0.6
Flask-Pydantic==0.9.0
pytest==7.0.1
